Fix bugs in models and change column name and association name
The first bug I fixed was to have the RecipeIngredient model inherit from ActiveRecord::Base. The other bugs that I fixed were to add this important association to the Recipe and Ingredient models: "has_many :recipe_ingredients". In the recipe_ingredients model, I changed :amount to :ingredient_amount for clarification.