Begin re-seeding database, then make a note about a bug with the #update method.
As noted in the NOTES.md file, #update worked as I expected when I initially re-seeded the database. However, when I tried to update the recipe_ingredients table's ingredient_amount attribute a second time (when its value was not nil), it changed the recipe and the recipe_ingredient, but NOT the ingredient.