Remove Ingredient model validation, and reset and reseed database
I don't need to validate the presence of every ingredient's name. An ingredient only needs a name if it's the first ingredient (taken care of in the recipes/new.erb file), or if the user specifies an amount and/or brand name.
That validation belongs to the recipe, not the ingredient, due to the schema structure.
It was necessary to reset and reseed the database because while testing the RecipesController, I mistakenly made a lot of recipes with nil users. This should be fixed in the next commit.