Remove name validation from Ingredient model
Because of the way I am creating recipes, I don't need a name validation in the Ingredient model. The idea is not to save a recipe unless it has valid ingredients.
The ingredients themselves won't be saved until I save the recipe.
I can add this Ingredient name validation back in, later on.