Refactor get '/recipes/:id/edit' route, add/edit comments, and change message in recipe edit form
The "get '/recipes/:id/edit'" route now first checks to see that the user is logged in before it looks for a recipe, and it gives an error message if the user is not logged in.
The route also has a separate action for when the recipe's user is not the current user, instead of redirecting to the recipe page when the user is not logged in and/or is not the recipe's author.
Finally, the recipe edit form now correctly greets the chef by his/her username.