Refactor code in recipe/show.erb page, move old code to NOTES.md, add comment to controller, and test the result
In the recipe/show.erb page, I deleted some obsolete code, and I modified and deleted a few comments for better clarification.
The comment in the controller added additional clarification to the edited portions of the recipe/show.erb page.
I tested the results by making a new recipe, this time with blank values for the ingredient's amount and brand name.