Add Ingredient names to recipe edit form, and delete unnecessary instance variable
@recipe_ingredients is unnecessary because I only need it on line 49 of the edit form. Therefore, @recipe.recipe_ingredients is a better option.
In order to list all of the ingredients by name in the edit form, I needed to make an instance variable for all of the ingredients.
The next commit will complete the Ingredients part of the form by adding blank Ingredient spaces until the number of ingredients to edit or create totals 5.