Finish recipe edit form by adding additional_ingredients, instructions, and notes
I ran into a problem with the <textarea> tag: it doesn't have a 'value' attribute that I can put the original user input into.
Also, the <input> tag does not have a type of 'textarea' or anything similar.
The solution was surprisingly simple: put the original user input between the opening and closing <textarea> tags!