Partially update recipe, add flash messages for success and failure, and test changes to code
The patch '/recipes/:id' route will currently update all of the recipe's params except the ingredients.
It correctly redirects to the recipe's show page with a flash message if the update was a success.

I do not know whether it redirects back to the recipe edit form with an error message for an invalid update.
The code for that is written, but I cannot test it out as I seem to have messed up Shotgun in my local environment; I hit Ctrl-Z in the terminal by mistake.