Enable and test autosave on a recipe's recipe_ingredients
This didn't work as I intended. I still haven't figured out how to update and save a recipe_ingredient, even with autosave enabled.
I could just call #update on the recipe_ingredients, but I don't want them to be updated until the recipe is saved, due to the validations in place.