Add #accepts_nested_attributes_for macro to Recipe model, use it in PATCH request, and update comments
The Recipe model now accepts nested attributes for recipe_ingredients (which was turned off by default).
By enabling that, I can now update my recipe with changes to (existing) recipe_ingredients, without first calling #update on the ingredients and recipe_ingredients.
Autosave alone could not do that. The recipe's validations are also correctly triggered (so far).
It was necessary to avoid updating the ingredients and recipe_ingredients before updating the recipe because of the recipe's complex validations.
If #update was called on a valid recipe_ingredient, but the recipe itself was invalid, then the changes to the recipe_ingredient would need to be reverted.