Update comments and code, attempt to complete recipe PATCH route, and test it
I decided to leave a few edge cases as stretch goals (letting the user remove an ingredient, changing the params hash, etc.)
My current goal is to allow the user to update ingredients and create new ingredients, and redirect with an error message when the ingredients are invalid.
Checking for valid non-ingredient recipe params is also an edge case/stretch goal, due to the 'edit' form's 'required' keywords.

This appears to have a few bugs in it, which will be resolved in the next commit. DO NOT REVERT TO THIS COMMIT!