Remove calls to #capitalize, fix bug in recipe show.erb file, and test the changes
In the recipe's 'post' and 'patch' routes, I chose to no longer use #capitalize on the brand names; capitalization is now up to the user.
The reason for the change: formerly, brand names such as 'Napa Valley' and 'Example Brand Name' would get changed to 'Napa valley' and 'Example brand name' when #capitalize was called.

This is the bug that I found in the recipe's show.erb file:
When two or more ingredients with the same name, but different amounts and/or brands, were saved and/or updated, the show.erb file would display the first ingredient multiple times, rather than the separate ingredients.
(To be clear, this only happened to ingredients with the same name; other ingredients with different names were still properly displayed.)
To fix this, I had the show.erb file iterate through the recipe's recipe_ingredients instead of its ingredients.