Change User login page to only show three types of messages, change flash messages, and add flash message
In an effort to simplify later CSS styling of flash messages, I have started limiting the flash message types.
In the case of user login, there are three types of flash messages: error, success, and login (the last one simply indicating whether the user is logged in or not.
This also required changing the controllers (flash[:logout] to flash[:success], e.g.)
Finally, I added a flash message to the login page for when users are not logged in; I may change this later, since the same message now shows up on two different pages.