Always redirect users to login page from logout route, and change flash[:logout] to flash[:success] and flash[:error]
As with other similar changes, changing flash[:logout] to flash[:success] and flash[:error] will make the flash messages easier to style with CSS.
This is also accounted for in the homepage (index.erb), as flash[:logout] no longer exists.
Lastly, I realized that when a user is logged out already and tries to access the logout route, it makes more sense to redirect to the login page than to the homepage.