Change flash[:message] to flash[:success], and test the changes.
This affects the patch and delete routes in the RecipesController, the routes that they redirect to, and the corresponding ERB files.
As with other changes, this will make the flash messages easier to style. This completes that set of changes.
From now on, I will be styling three types of flash messages: login, error, and success.