Add minor style change, stretch goal/idea, and bug fix idea to NOTES.md
The important part here is the bug fix: I discovered a bug in the recipe's new and update forms that allows users to add the same ingredient multiple times to a recipe.
This issue has turned out to be more complicated than it first appeared, as noted in the NOTES.md file.