Adjust Markdown in README, and add development note
I finally figured out and implemented the correct way to make an unordered sublist item with Markdown.
I also realized that one of the ordered list items in the Installation section is better as an unordered sublist item.
The new development note addresses a DELETE feature that I added through ActiveRecord.