Find the best size of an image to add to a recipe by adding/deleting images of various sizes
The best (maximum) size of an image turned out to be 480 px by 320 px.
This commit technically should have been made before the previous commit, since the changes to the index file and CSS file were made AFTER I experimented with image sizes.
However, since the images in question were deleted, the order of commits (in this case) isn't a huge deal.