Replace Image URL note with (styled) directions to home page
This commit, along with the recent commit titled 'Edit Recipe Rolodex description, and add and style instructions for giving a recipe an image', finishes moving the Image URL instructions from the recipe 'create' and 'edit' forms to the home page, and expanding them there.
Since the home page also serves as an About page, I thought it best to move and expand the instructions there, rather than cluttering up the recipe forms.