Add #all_sorted class method to Recipe model
Originally, I had sorted the recipes alphabetically by downcased name and username from the recipe index page.
This logic has been moved to the Recipe model.
As will be shown in the next commit, this class method is now called from the RecipesController when setting up the @recipes instance variable.