Start refactoring the RecipesController, edit error message, and add stretch goal
Some of the logic in the "get '/recipes'" route was replaced by calling the ApplicationController's #redirect_if_not_logged_in helper method.
Also in that route, the recipes are now sorted with the Recipe model's #all_sorted class method before being sent to the recipe index page as @recipes.
Similar refactoring will likely be applied to other routes.
The edge case in the "post '/recipes'" route turned out to be easier to reach than I realized (thanks to the browser development tools). The error message was updated to reflect this.
Finally, I added/modified a stretch goal, with the intent of refactoring the params hash to make the recipe PATCH route much simpler.