Refactor recipe index page by removing an edge case and moving code to the controller and model
The edge case in the index page redirected users with an error message if they were not logged in.
That code belongs in the RecipesController and is thus not needed here.
I originally checked Recipe.all in the index file; however, that made an unnecessary call to the recipes database, since all of the recipes were already saved into @recipes in the RecipesController.
Finally, I had originally sorted the recipes from this index page; that logic has now correctly been moved to the Recipe model and called from the RecipesController.