Refactor #redirect_if_not_logged_in to accept a default argument
I realized that the error message 'You must be logged in to view this page.' doesn't always make sense.
This method can also be called when a logged out user tries to create, edit, or delete a recipe, not necessarily view a page.
The default message is still 'You must be logged in to view this page.', but I can now change it with a custom message such as '...to create a recipe.'