Add old comments and code, delete old stretch goal, and add new stretch goal
I added some old comments and code from a route in the RecipesController to have on hand for future reference.
The stretch goal that I deleted has already been implemented.
The stretch goal that I added concerns inline styling and may not be possible to do without advanced HTML and CSS knowledge.