Refactor multiple routes in RecipesController with the #redirect_if_not_logged_in helper method, and update comment
Most, if not all, routes in the RecipesController check to see whether a user is logged in.
They redirect with an error message if the user is not logged in.
This has been refactored into a #redirect_if_not_logged_in helper method, which is in the ApplicationController.
Although these are edge cases, some have been marked as important, because the program will break if a user tries to create, update, or delete a recipe while logged out.
One of the comments referred to a no-longer-existing comment in the recipe show file; it has been updated accordingly.