Refactor and rename Recipe class method and update corresponding route in RecipesController
I realized that my program doesn't just sort all of the recipes; it also sorts recipes that belong to a user.
Both sets of recipes are sorted with the same code.
So, I renamed the #Recipe.all_sorted class method as #Recipe.sort_recipes, and I made it accept a recipes argument.
This can now be used by multiple controllers; feasibly, if I wanted to list and sort the recipes that include a certain ingredient, then I could use this method for that as well.