Edit/delete comments, add a helper method, and use it to refactor a route
Some comments were either unnecessary, or they needed to be made more concise.
The helper method is #redirect_if_unauthorized_user_tries_to(action).
It was used to refactor the "get '/recipes/:id/edit'" route to prevent the wrong user from accessing the recipe edit form.
It will also be used to refactor other routes.
Since it specifically relates to recipes, this helper method is in the RecipesController, rather than the ApplicationController.