Rename and change a helper method, and update recipe routes that are affected by it
I changed the name of the #redirect_if_recipe_does_not_exist helper method to #redirect_if_nonexistent.
It now also accepts an object as an argument.
For now, it will be passed instance variables and local variables until I refactor it.
My goal is to move this to the ApplicationController so that it can be used on both recipes and users.