Change helper method, update affected routes, and test the changes
I changed the ApplicationController helper method #redirect_if_nonexistent to accept an object type in addition to an object.
It also displays the object type (not just "recipes") in its error message.
The intent here is to make this method usable by the UsersController, not just the RecipesController.
For now, it redirects to the '/recipes' route, but this will change in another commit.
I also had to update the affected routes in the RecipesController and make sure it all works.