Add helper method, and use it to refactor a UsersController route
I added a #redirect_if_not_existing helper method to the ApplicationController.
I used it to redirect users to the homepage if it could not find a chef, though it can be used for any object of any type.
I then refactored the UsersController's 'get /users/:username' route with #redirect_if_not_existing.
It is intentionally similar to the helper method above it, because both methods will be combined into one in another commit.