Add comment, modify comment, and refactor route with helper method
The comment that I added, marks the end of the UsersController (similar comments exist in the other controllers).
I then refactored the 'post /users/login' route to use the #redirect_if_not_existing helper method.
As such, I had to update a comment in that route to reflect the change.