Change redirect location in helper method, delete corresponding note, and add important comment
The #redirect_if_unauthorized_user_tries_to helper method originally redirected unauthorized users to the display page of the recipe that they tried to edit or delete.
I realized that this was a bad idea, particularly if the hacker didn't know that the recipe existed and simply typed a random recipe id into the URL.
For this reason, I redirected them to the list of recipes instead and deleted the corresponding idea from NOTES.md.
Lastly, I realized that, this helper method will break if it is called while the user is logged out (unlikely though that may be); as such, I added a note and will fix this with more refactoring later.